fix warning with -Wconversion in Agent & Environment

+++ src/environment/Environment.cpp
@@ -48,7 +48,7 @@ void Environment::addSensor(const std::string& name, const SensorPtr& sensor)
 
 int Environment::getNumSensor() const
 {
-    return sensor_list.size();
+    return static_cast<int>(sensor_list.size());
 }
 
 const SensorPtr Environment::getSensorByName(const std::string& name) const
